isolate tests to support changes in go versions
