+++ bridge_test.go
@@ -1,12 +1,9 @@
 package huego
 
 import (
-	"context"
 	"fmt"
 	"strings"
 	"testing"
-
-	"github.com/jarcoal/httpmock"
 )
 
 func ExampleBridge_CreateUser() {
@@ -60,59 +57,5 @@ func TestUpdateBridgeConfigError(t *testing.T) {
 	_, err := b.GetConfig()
 	if err == nil {
 		t.Fatal("Expected error not to be nil")
-	}
-}
-
-func TestBridge_getAPIPathError(t *testing.T) {
-	b := New("invalid hostname", "")
-	expectedOlderGoVersion := "parse http://invalid hostname: invalid character \" \" in host name"
-	expectedNewerGoVersion := "parse \"http://invalid hostname\": invalid character \" \" in host name"
-	_, err := b.getAPIPath("/")
-	if errString := err.Error(); errString != expectedOlderGoVersion && errString != expectedNewerGoVersion {
-		t.Fatalf("Expected error %s but got %s", expectedOlderGoVersion, errString)
-	}
-}
-
-func TestBridge_getError(t *testing.T) {
-	httpmock.Deactivate()
-	defer httpmock.Activate()
-	expectedOlderGoVersion := "Get invalid%20hostname: unsupported protocol scheme \"\""
-	expectedNewerGoVersion := "Get \"invalid%20hostname\": unsupported protocol scheme \"\""
-	_, err := get(context.Background(), "invalid hostname")
-	if errString := err.Error(); errString != expectedOlderGoVersion && errString != expectedNewerGoVersion {
-		t.Fatalf("Expected error %s but got %s", expectedOlderGoVersion, errString)
-	}
-}
-
-func TestBridge_putError(t *testing.T) {
-	httpmock.Deactivate()
-	defer httpmock.Activate()
-	expectedOlderGoVersion := "Put invalid%20hostname: unsupported protocol scheme \"\""
-	expectedNewerGoVersion := "Put \"invalid%20hostname\": unsupported protocol scheme \"\""
-	_, err := put(context.Background(), "invalid hostname", []byte("huego"))
-	if errString := err.Error(); errString != expectedOlderGoVersion && errString != expectedNewerGoVersion {
-		t.Fatalf("Expected error %s but got %s", expectedOlderGoVersion, errString)
-	}
-}
-
-func TestBridge_postError(t *testing.T) {
-	httpmock.Deactivate()
-	defer httpmock.Activate()
-	expectedOlderGoVersion := "Post invalid%20hostname: unsupported protocol scheme \"\""
-	expectedNewerGoVersion := "Post \"invalid%20hostname\": unsupported protocol scheme \"\""
-	_, err := post(context.Background(), "invalid hostname", []byte("huego"))
-	if errString := err.Error(); errString != expectedOlderGoVersion && errString != expectedNewerGoVersion {
-		t.Fatalf("Expected error %s but got %s", expectedOlderGoVersion, errString)
-	}
-}
-
-func TestBridge_deleteError(t *testing.T) {
-	httpmock.Deactivate()
-	defer httpmock.Activate()
-	expectedOlderGoVersion := "Delete invalid%20hostname: unsupported protocol scheme \"\" "
-	expectedNewerGoVersion := "Delete \"invalid%20hostname\": unsupported protocol scheme \"\""
-	_, err := delete(context.Background(), "invalid hostname")
-	if errString := err.Error(); errString != expectedOlderGoVersion && errString != expectedNewerGoVersion {
-		t.Fatalf("Expected error %s but got %s", expectedOlderGoVersion, errString)
 	}
 }
